backend/fastapi/pyproject.toml
[tool.poetry]
name = "electronic-health-record"
version = "0.1.0"
description = ""
authors = ["ingeniom22 <jamesmichael0444@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
torch = "^2.2.2"
torchvision = "^0.17.2"
neural-compressor = "^2.5"
pymongo = "^4.6.3"
motor = "^3.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
